新增VPN验证码/顶号功能。
