+++ Views/login.xaml
@@ -90,6 +90,7 @@
                 <RowDefinition Height="Auto"></RowDefinition>
                 <RowDefinition Height="Auto"></RowDefinition>
                 <RowDefinition Height="Auto"></RowDefinition>
+                <RowDefinition Height="Auto"></RowDefinition>
             </Grid.RowDefinitions>
 
             <StackPanel Grid.Row="0"  Orientation="Horizontal" Spacing="5">
@@ -102,13 +103,14 @@
             <TextBlock Grid.Row="1" x:Name="de" Text="VPN的凭据通常与浙大通行证一致。" Foreground="{ThemeResource TextFillColorSecondaryBrush}" FontSize="12" Margin="0,10,0,0"></TextBlock>
             <TextBox Grid.Row="2" Header="账号" x:Name="idbox" Margin="0,10,0,0" ></TextBox>
             <PasswordBox Grid.Row="3" Header="密码" x:Name="passbox" ></PasswordBox>
-            <dev:ProgressButton Grid.Row="4" HorizontalAlignment="Right" Margin="0,10,0,0" x:Name="Link" IsChecked="False" Click="Link_Click" ProgressForeground="#60d2b3ae"  CheckedContent="登录中···" IsIndeterminate="True" ProgressRingVisibility="Collapsed">
+            <TextBox x:Name="captchabox" Grid.Row="4" Header="验证码" Text="" TextChanged="captchabox_TextChanged" MaxLength="4" Visibility="Collapsed"></TextBox>
+            <Image x:Name="captcha" Grid.Row="5" HorizontalAlignment="Left" Height="40" Stretch="Uniform"></Image>
+            <dev:ProgressButton Grid.Row="5" HorizontalAlignment="Right" Margin="0,10,0,0" x:Name="Link" IsChecked="False" Click="Link_Click" ProgressForeground="#60d2b3ae"  CheckedContent="登录中···" IsIndeterminate="True" ProgressRingVisibility="Collapsed">
                 <StackPanel Orientation="Horizontal" Spacing="5">
                     <ic:SymbolIcon Symbol="ArrowEnter" FontSize="14" ></ic:SymbolIcon>
                     <TextBlock Text="登录" ></TextBlock>
                 </StackPanel>
             </dev:ProgressButton>
-           
         </Grid>
         <Grid x:Name="LoginPane"  Grid.Column="2" Grid.Row="1" Padding="20" RowSpacing="10">
             <Grid.RowDefinitions>
